+++ kotti_grid/tests/test_views.py
@@ -4,4 +4,4 @@ def test_gridbrowser(db_session, dummy_request, root):
     from kotti_grid.views import GridBrowser
     gt = GridBrowser(root, dummy_request)
     browser = gt.gridbrowser()
-    assert browser == {}
+    assert browser['form'].startswith('<form')