+++ kotti_grid/views.py
@@ -1,3 +1,7 @@
+import colander
+import deform
+from deform import Form
+
 from pyramid.view import view_config
 from pyramid.view import view_defaults
 
@@ -20,6 +24,28 @@ def save_grid(context, request):
     return save_grid_settings(tiles, 'tiles')
 
 
+use_values = (('use_title', _(u'Use title')),
+              ('use_description', _(u'Use description')),
+              ('use_title_and_description', _(u'Use title and description')),
+              ('use_body_text', _(u'Use body text')),
+              ('use_image', _(u'Use image')),
+              ('use_internal_image', _(u'Use internal image')),)
+
+
+class UseSchemaNode(colander.SchemaNode):
+    name = 'use'
+    title = _(u'Use')
+    missing = u'use_title',
+    widget = deform.widget.SelectWidget(values=use_values)
+
+
+class GridSchema(colander.MappingSchema):
+    use = UseSchemaNode(colander.String())
+    extra_style = colander.SchemaNode(colander.String(),
+                                 title=_(u'Extra styles'),
+                                 missing=u'',)
+
+
 @view_defaults(context=IContent,
                request_method="GET")
 class GridBrowser():
@@ -32,7 +58,9 @@ class GridBrowser():
     @view_config(name="gridbrowser",
                  renderer="kotti_grid:templates/browser.pt")
     def gridbrowser(self):
-        return {}
+        schema = GridSchema()
+        form = Form(schema, buttons=())
+        return {"form": form.render()}
 
 
 def includeme(config):