+++ CHANGES.rst
@@ -7,6 +7,7 @@ Changelog
 * Added settings for the grid.
 * Move initialization of the grid to the template.
 * Added some documentation in the README.
+* Added new form settings to the gridbrowser.
 
 
 0.1a1